api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
supabase==1.0.4
python-dotenv==1.0.0
stripe==5.5.0
requests==2.31.0
Pillow==10.0.0
PyPDF2==3.0.1
reportlab==4.0.4
gunicorn==21.2.0
psycopg2-binary==2.9.7